feat: 설정 페이지 디자인 수정
